Refactored file structure and added url-loader to webpack.
In order to have it similar to webviz-subsurface-components.
