+++ src/lib/components/WebvizPluginPlaceholder/utils/WebvizContentOverlay.js
