+++ src/lib/components/WebvizPluginPlaceholder/utils/WebvizToolbarButton.js
