+++ src/lib/components/WebvizPluginPlaceholder/utils/download_file.js
