+++ src/lib/index.js
@@ -1,7 +1,7 @@
 /* eslint-disable import/prefer-default-export */
-import WebvizPluginPlaceholder from "./components/WebvizPluginPlaceholder.react";
-import ColorScales from "./components/ColorScales.react";
-import Select from "./components/Select.react";
+import WebvizPluginPlaceholder from "./components/WebvizPluginPlaceholder";
+import ColorScales from "./components/ColorScales";
+import Select from "./components/Select";
 
 import "./flexbox.css"
